+++ tests/testthat/test-initialization.R
@@ -60,19 +60,19 @@ test_that("zoom-in search works", {
   data <- create_test_data(model)
   
   sim_cache <- create_sim_cache()
-  par <- do_zoom_in_search(model, data, 1, sim = 20, cores = 1, sim_cache)
+  par <- do_zoom_in_search(model, data, 1, sim = 20, cores = 1, sim_cache, 0.05)
   expect_that(par, is_a("matrix"))
   expect_equal(dim(par), c(1, model$get_par_number()))
   expect_true(all(par >= 0 & par <= 1))
   
   sim_cache <- create_sim_cache()
-  par <- do_zoom_in_search(model, data, 2, sim = 20, cores = 1, sim_cache)
+  par <- do_zoom_in_search(model, data, 2, sim = 20, cores = 1, sim_cache, 0.1)
   expect_that(par, is_a("matrix"))
   expect_equal(dim(par), c(2, model$get_par_number()))
   expect_true(all(par >= 0 & par <= 1))
   
   sim_cache <- create_sim_cache()
-  par <- do_zoom_in_search(model, data, 3, sim = 20, cores = 1, sim_cache)
+  par <- do_zoom_in_search(model, data, 3, sim = 20, cores = 1, sim_cache, 0.2)
   expect_that(par, is_a("matrix"))
   expect_equal(dim(par), c(3, model$get_par_number()))
   expect_true(all(par >= 0 & par <= 1))
@@ -92,20 +92,37 @@ test_that("getting the start positions works", {
   expect_equal(get_start_pos(model, data, 3, 20, "middle", 1, sim_cache),
                matrix(0.5, 3, model$get_par_number()))
   
+  # random
+  pos <- get_start_pos(model, data, 1, 20, "random", 1, sim_cache)
+  expect_equal(dim(pos), c(1, 2))
+  expect_equal(length(unique(pos)), 2)
+  expect_true(all(pos >= 0 & pos <= 1))
+  
+  pos <- get_start_pos(model, data, 2, 20, "random", 1, sim_cache)
+  expect_equal(dim(pos), c(2, 2))
+  expect_equal(length(unique(pos)), 4)
+  expect_true(all(pos >= 0 & pos <= 1))
+  
+  pos <- get_start_pos(model, data, 3, 20, "random", 1, sim_cache)
+  expect_equal(dim(pos), c(3, 2))
+  expect_equal(length(unique(pos)), 6)
+  expect_true(all(pos >= 0 & pos <= 1))
+  
+  
   # initial search
   sim_cache <- create_sim_cache()
-  pos <- get_start_pos(model, data, 1, 20, "initial-search", 1, sim_cache)
+  pos <- get_start_pos(model, data, 1, 20, "initial-search", 1, sim_cache, 0.05)
   expect_that(pos, is_a("matrix"))
   expect_true(all(pos >= 0 & pos <= 1))
   
   # zoom-in
   sim_cache <- create_sim_cache()
-  pos <- get_start_pos(model, data, 1, 20, "zoom-in", 1, sim_cache)
+  pos <- get_start_pos(model, data, 1, 20, "zoom-in", 1, sim_cache, 0.05)
   expect_that(pos, is_a("matrix"))
   expect_true(all(pos >= 0 & pos <= 1))
   
   # errors
   sim_cache <- create_sim_cache()
-  expect_error(get_start_pos(model, data, 1, 20, "1", 1, sim_cache))
-  expect_error(get_start_pos(model, data, 1, 20, 1, 1, sim_cache))
+  expect_error(get_start_pos(model, data, 1, 20, "1", 1, sim_cache, 0.05))
+  expect_error(get_start_pos(model, data, 1, 20, 1, 1, sim_cache, 0.1))
 })